Move windows_subsystem field from CodegenResults to CrateInfo
